Add ansible parameter to uninstall
The new v3.0.4 tag was added, but it was edited after.
This raises an error during an update of existing git folder.
The solution is to add an option to delete existing checkout code.
With this we can deploy with `--tags pakiti -e pakiti_uninstall=true`
and it will delete and checkout again a fresh pakiti.
Resolves #5